refactor(virtual-scroll): refactor virtualScroll component
