+++ src/components/StrengthMeter/index.ts
@@ -1,1 +1,4 @@
-export { default as StrengthMeter } from './src/StrengthMeter.vue';
+import { withInstall } from '/@/utils';
+import strengthMeter from './src/StrengthMeter.vue';
+
+export const StrengthMeter = withInstall(strengthMeter);
