+++ src/components/Time/index.ts
@@ -1,1 +1,4 @@
-export { default as Time } from './src/Time.vue';
+import { withInstall } from '/@/utils/index';
+import time from './src/Time.vue';
+
+export const Time = withInstall(time);
