+++ src/components/Tinymce/index.ts
@@ -1,2 +1,4 @@
-import Tinymce from './src/Editor.vue';
-export { Tinymce };
+import { withInstall } from '/@/utils/index';
+import tinymce from './src/Editor.vue';
+
+export const Tinymce = withInstall(tinymce);
