--- src/components/VirtualScroll/src/index.less
@@ -1,18 +1,0 @@
-.virtual-scroll {
-  position: relative;
-  display: block;
-  width: 100%;
-  max-width: 100%;
-  overflow: auto;
-  flex: 1 1 auto;
-
-  &__container {
-    display: block;
-  }
-
-  &__item {
-    position: absolute;
-    right: 0;
-    left: 0;
-  }
-}
